Do a major refactoring of the flow exporter into two apps
This separates the metering code from the export code, which
is a cleaner separation. Currently the flow expiry code is in
the export app, but it's not clear that that's the right place.

It may be more appropriate to put it in the metering app, with
expired packets going in some kind of queue perhaps in the flow
cache for the other app to read out of.
